requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.5.1
Automat==22.10.0
Babel==2.13.1
backoff==2.2.1
backports.functools-lru-cache==1.6.5
bcrypt==4.0.1
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
cattrs==23.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
constantly==15.1.0
contourpy==1.2.0
coverage==7.4.1
cryptography==41.0.5
cssselect==1.2.0
cycler==0.12.1
dataclasses==0.8
dataclasses-json==0.5.7
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
dependency-injector==4.41.0
Deprecated==1.2.14
diffusers==0.24.0
dill==0.3.7
distro==1.8.0
docutils==0.20.1
ecdsa==0.18.0
esbonio==0.16.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fake-useragent==1.3.0
Faker==20.0.3
fastapi==0.104.1
fastapi-simple-security==1.3.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
gekko==1.0.6
gitdb==4.0.11
GitPython==3.1.40
gmpy2==2.1.2
google-api-core==2.13.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.20.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jq==1.6.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langserve==0.0.39
langsmith==0.0.84
lsprotocol==2023.0.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-service==2.4.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.16.2
openai==0.28.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
optimum==1.14.0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.2
parsel==1.8.1
parso==0.8.3
passlib==1.7.4
peft==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
platformdirs==3.11.0
pluggy==1.3.0
posthog==3.0.2
prompt-toolkit==3.0.40
Protego==0.3.0
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.1
pydeck==0.8.1b0
PyDispatcher==2.0.5
pygls==1.2.0
Pygments==2.16.1
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyPika==0.48.9
PySocks==1.7.1
pyspellchecker==0.7.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-stub==1.1.0
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.7.0
rouge==1.0.1
rpds-py==0.13.2
rsa==4.9
sacremoses==0.0.53
safetensors==0.3.3
scikit-learn==1.3.2
scipy==1.11.3
Scrapy==2.11.0
scrapy-fake-useragent==1.4.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==18.1.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.23
sse-starlette==1.8.2
stack-data==0.6.2
starlette==0.27.0
streamlit==1.30.0
stringcase==1.2.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tldextract==5.1.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.15.2a0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.33.3
triton==2.1.0
Twisted==22.10.0
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
w3lib==2.1.2
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.9
websocket-client==1.6.4
websockets==12.0
wheel==0.41.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1

--extra-index-url=https://download.pytorch.org/whl/cpu